requirements.txt
aiohttp==3.6.2
aiohttp-jinja2==1.2.0
aiohttp-jwt==0.5.0
argon2-cffi==19.2.0
async-timeout==3.0.1
asyncpg==0.20.1
attrs==19.3.0
cffi==1.14.0
chardet==3.0.4
envparse==0.2.0
idna==2.9
Jinja2==2.11.1
MarkupSafe==1.1.1
marshmallow==3.5.1
multidict==4.7.5
psycopg2-binary==2.8.4
pycparser==2.20
PyJWT==1.7.1
six==1.14.0
SQLAlchemy==1.3.13
uvloop==0.14.0
yarl==1.4.2

crypto~=1.4.1